Add styling to the quote/ cards sections
The styling is done on a 50vh/ 50vh basis, where
they both together take up 100 % of user's screen
In the cards section, the cards are styled so to
take up to 3 more. The quotation can be expanded
too.

Small HTML changes here and there to accommodate
for the styling, such as the viewport sizes and
similar. Small stylish HTML changes, as well